@@ -1,4 +1,4 @@
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
@@ -25,46 +25,40 @@ import { MatSelect } from '@angular/material/select';
 import { MatFormField } from '@angular/material/form-field';
 import { MdbRippleModule } from "mdb-angular-ui-kit/ripple";
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    HeaderComponent,
-    FooterComponent,
-    AboutComponent,
-    BoredApiPageComponent,
-    HomeComponent,
-    ProjectsComponent,
-    ProjectModalComponent,
-    ProjectModalEditComponent,
-    ProjectModalNewComponent,
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    ReactiveFormsModule,
-    AppRoutingModule,
-    MatProgressSpinner,
-    BrowserAnimationsModule,
-    MatProgressSpinner,
-    MatOption,
-    MatSelect,
-    MatOption,
-    MatFormField,
-    MdbRippleModule,
-  ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useFactory: (router: Router) => {
-        return new AuthInterceptor(router);
-      },
-      multi: true,
-      deps: [Router],
-    },
-    AuthGuard,
-    AuthService,
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        LoginComponent,
+        HeaderComponent,
+        FooterComponent,
+        AboutComponent,
+        BoredApiPageComponent,
+        HomeComponent,
+        ProjectsComponent,
+        ProjectModalComponent,
+        ProjectModalEditComponent,
+        ProjectModalNewComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        ReactiveFormsModule,
+        AppRoutingModule,
+        MatProgressSpinner,
+        BrowserAnimationsModule,
+        MatProgressSpinner,
+        MatOption,
+        MatSelect,
+        MatOption,
+        MatFormField,
+        MdbRippleModule], providers: [
+        {
+            provide: HTTP_INTERCEPTORS,
+            useFactory: (router: Router) => {
+                return new AuthInterceptor(router);
+            },
+            multi: true,
+            deps: [Router],
+        },
+        AuthGuard,
+        AuthService,
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {}
