@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProjectModalEditComponent } from './project-modal-edit.component';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { FormBuilder } from '@angular/forms';
 
@@ -11,13 +11,11 @@ describe('ProjectModalEditComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [
-        ProjectModalEditComponent,
-        HttpClientModule,
+    imports: [ProjectModalEditComponent,
         NgbActiveModal,
-        FormBuilder,
-      ],
-    }).compileComponents();
+        FormBuilder],
+    providers: [provideHttpClient(withInterceptorsFromDi())]
+}).compileComponents();
 
     fixture = TestBed.createComponent(ProjectModalEditComponent);
     component = fixture.componentInstance;
