@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProjectModalNewComponent } from './project-modal-new.component';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 describe('ProjectModalNewComponent', () => {
@@ -10,8 +10,9 @@ describe('ProjectModalNewComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [ProjectModalNewComponent, HttpClientModule, NgbActiveModal],
-    }).compileComponents();
+    imports: [ProjectModalNewComponent, NgbActiveModal],
+    providers: [provideHttpClient(withInterceptorsFromDi())]
+}).compileComponents();
 
     fixture = TestBed.createComponent(ProjectModalNewComponent);
     component = fixture.componentInstance;
